+++ e2e/testdata/fn-eval/selectors/basicpipeline/.krmignore
@@ -1,0 +1,1 @@
+.expected
