+++ internal/cmdrender/executor_test.go
@@ -18,11 +18,9 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/GoogleContainerTools/kpt/internal/types"
-	v1 "github.com/GoogleContainerTools/kpt/pkg/api/kptfile/v1"
+	"github.com/GoogleContainerTools/kpt/internal/fnruntime"
 	"github.com/stretchr/testify/assert"
 	"sigs.k8s.io/kustomize/kyaml/kio"
-	"sigs.k8s.io/kustomize/kyaml/yaml"
 )
 
 func TestPathRelToRoot(t *testing.T) {
@@ -222,148 +220,8 @@ metadata:
 			assert.NoError(t, err)
 			input, err := kio.ParseAll(tc.input)
 			assert.NoError(t, err)
-			result := mergeWithInput(output, selectedInput, input)
+			result := fnruntime.MergeWithInput(output, selectedInput, input)
 			actual, err := kio.StringAll(result)
-			assert.NoError(t, err)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
-
-func TestIsMatch(t *testing.T) {
-	tests := []struct {
-		name     string
-		input    string
-		selector v1.Selector
-		expected bool
-	}{
-		{
-			name: "kind match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				Kind: "Deployment",
-			},
-			expected: true,
-		},
-		{
-			name: "name match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				Name: "nginx-deployment",
-			},
-			expected: true,
-		},
-		{
-			name: "namespace match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  namespace: staging
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				Namespace: "staging",
-			},
-			expected: true,
-		},
-		{
-			name: "apiVersion match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				APIVersion: "apps/v1",
-			},
-			expected: true,
-		},
-		{
-			name: "GVKNN match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  namespace: staging
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				Name:       "nginx-deployment",
-				Namespace:  "staging",
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			expected: true,
-		},
-		{
-			name: "namespace not matched but rest did",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  namespace: staging
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				Name:       "nginx-deployment",
-				Namespace:  "prod",
-				Kind:       "Deployment",
-				APIVersion: "apps/v1",
-			},
-			expected: false,
-		},
-		{
-			name: "packagePath match",
-			input: `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: nginx-deployment
-  annotations:
-    internal.config.k8s.io/kpt-resource-id: "0"
-    internal.config.kubernetes.io/package-path: "/path/to/root/pkg/db"
-spec:
-  replicas: 3`,
-			selector: v1.Selector{
-				PackagePath: "./db",
-			},
-			expected: true,
-		},
-	}
-
-	for i := range tests {
-		tc := tests[i]
-		t.Run(tc.name, func(t *testing.T) {
-			node, err := yaml.Parse(tc.input)
-			assert.NoError(t, err)
-			var rootPackagePath string
-			if tc.selector.PackagePath != "" {
-				rootPackagePath = "/path/to/root/pkg"
-			}
-			actual, err := isMatch(node, tc.selector, &selectionContext{rootPackagePath: types.UniquePath(rootPackagePath)})
 			assert.NoError(t, err)
 			assert.Equal(t, tc.expected, actual)
 		})
